Apply the PR 3 review suggestions
Types and seams:
- process: `Group`, returned by `spawn_detached`, is the only thing a
  group kill takes, so a capture child that shares our group can no
  longer be group-killed; kill_group is gone.
- timer: `Plan::from_choices` holds the Timer dialog's rules (positive
  length, next occurrence, stop after the start) with a `Refused` reason;
  the GTK dialog shrinks to one call. `first_reading` picks the earlier
  instant of a repeated hour by timestamp: chrono 0.4.45 lists the later
  one first for Local on macOS, which the new DST test (a child test
  process pinned to America/New_York) caught.
- playback: the output comes from `momr_platform::playback::OUTPUT`
  (None off macOS), args are a pure tested function, empty playlists are
  refused and negative starts clamped, `ended` waits for the stderr
  reader instead of a fixed sleep, `probe_duration_us` returns Option.
- One clock format, `timer::clock`, replaces playback::clock and
  ui::format_elapsed; core CLIs return u8 exit codes.
- Socket commands skip the `defaults` language lookup unless they print.

Tests: process groups (leader, grandchild, SIGTERM, already gone), the
socket takeover of a stale file and refusal of a live one, safe_name,
playback arguments, Plan::from_choices, DST in New York; AppKit test
target (resume appends, peak of both channels, meter scale).

AppKit shell: `Source` enum, a recorder state enum with `stopped`
terminal, MeterView clamps in didSet and runs its timer only in a window.

Verified: cargo fmt/clippy/test (120 Rust tests), swift test (3), the
AppKit binary starts both captures, `momr stop` without an app.

+++ crates/momr-core/src/diarize.rs
@@ -200,7 +200,7 @@ pub fn turn_start_near(turns: &[Turn], speaker: usize, around_ms: i64) -> Option
 
 /// `diarize <audio> [--speakers N]`: prints the speaker turns as JSON, for
 /// comparing diarization engines on the same file.
-pub fn cli(args: &[String]) -> i32 {
+pub fn cli(args: &[String]) -> u8 {
     let mut path = None;
     let mut speakers = None;
     let mut iter = args.iter();
